# lab7/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Access API",
    "version": "1.0.0"
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/token": {
      "post": {
        "description": "Returns a token for the user to access the API",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Course",
            "in": "body",
            "description": "JSON data for the course",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token created successfully"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "description": "Returns a course with the given ID",
        "consumes": [
          "id"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the course to return",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "put": {
        "description": "Updates a course with the given ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the course to update",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Course",
            "in": "body",
            "description": "JSON data for the course",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Courses"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Course updated successfully"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      },
      "delete": {
        "description": "Delete a course with the given ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the course to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted successfully"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    },
    "/api/courses": {
      "get": {
        "description": "Returns the list of all courses with pagination support",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "required": false,
            "type": "integer"
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of courses per page (default: 4)",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of courses with pagination metadata"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      },
      "post": {
        "description": "Add a course",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Course",
            "in": "body",
            "description": "JSON data for adding a course",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Courses"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Course added successfully"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    }
  },
  "definitions": {
    "Courses": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "level": {
          "type": "integer"
        },
        "description2": {
          "type": "string"
        },
        "description3": {
          "type": "string"
        }
      }
    },
    "Token": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}